move destination implementation into destination package
